Refactor redeploy to separate service part
